[WIP] Server ordering, client paginated array, __table refactor Added in the basics for server-side ordering based on params.  The implementation is similar to the totem_filter approach.
Implemented a server/client version of the __table component to allow
for server-side or client-side sorting, ordering, and filtering.
Currently, the server-side is not fully integrated for the heading
sorts, but the pagination is functional.  The client-side is mostly
function for now.  The pagination component needs to be restyled
heavily.

Added a client paginated array that mirrors the API for the server
paginated array at the Totem level.  This keeps it consistent for things
like `get_next_page` in order to support compatibility for the __table
component and others.