[FIX] require test helper method
This method in the casespace base helper was not grabbing helper files
that were not within its own engine. Now there is an options hash
included with the method with the key of [:engine] that can be passed in
when used in test files outside of the casespace engine.

e.g. in the readiness_assurance_helper.rb

require_test_helper_files('../readiness_assurance/helpers',
engine: 'thinkspace-readiness-assurance')